notification skeleton implemented

+++ src/main/java/application/repository/RoleRepository.java
@@ -1,10 +1,10 @@
 package application.repository;
 
-import application.model.Account.Role;
+import application.model.account.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RoleRepository extends JpaRepository<Role, Integer> {
+public interface RoleRepository extends JpaRepository<Role, Long> {
 	Role findByRole(String role);
 }
